Retain the behaviour of format in numbers.format_currency
b449a16 introduced an API change by changing the behaviour of the format
param, instead this commit adds a format_type param and the documentation
and tests to accompany it.

Note that currency_formats returns a NumberPattern already, so there is no
need to call parse_pattern on the value returned.